update: rust-vm - clean mods
