+++ snoozetime.github.io/rust-evm/src/lib.rs
@@ -1,0 +1,3 @@
+mod evm;
+
+pub use crate::evm::{opcode, vm};
